Separate errors.py into errors/self.py and errors/peer.py
Move self_api.py and peer_api.py into an api subpackage

Split serializers.py into serializers/base.py and serializers/peer.py
