+++ apps/openassessment/assessment/errors/__init__.py
@@ -1,0 +1,8 @@
+"""
+Export errors from all modules defined in this package.
+"""
+
+# pylint:disable=W0401
+
+from .peer import *
+from .self import *
