+++ apps/openassessment/assessment/errors/peer.py
@@ -1,5 +1,5 @@
 """
-Errors for the assessment app.
+Errors for the peer assessment.
 """
 import copy
 
@@ -23,7 +23,7 @@ class PeerAssessmentRequestError(PeerAssessmentError):
     """
 
     def __init__(self, field_errors):
-        Exception.__init__(self, repr(field_errors))
+        super(PeerAssessmentRequestError, self).__init__(repr(field_errors))
         self.field_errors = copy.deepcopy(field_errors)
 
 
@@ -45,28 +45,3 @@ class PeerAssessmentInternalError(PeerAssessmentError):
 
     """
     pass
-
-
-class SelfAssessmentError(Exception):
-    """Generic Self Assessment Error
-
-    Raised when an error occurs while processing a request related to the
-    Self Assessment Workflow.
-
-    """
-    pass
-
-
-class SelfAssessmentRequestError(SelfAssessmentError):
-    """
-    There was a problem with the request for a self-assessment.
-    """
-    pass
-
-
-class SelfAssessmentInternalError(SelfAssessmentError):
-    """
-    There was an internal problem while accessing the self-assessment api.
-    """
-    pass
-
