+++ apps/openassessment/assessment/serializers/__init__.py
@@ -1,0 +1,8 @@
+"""
+Export serializers from each module in this package.
+"""
+
+# pylint: disable=W0401
+
+from .base import *
+from .peer import *
